+++ desktop/src/main/log.ts
@@ -84,6 +84,50 @@ const logDebug = (param: () => unknown) => {
 };
 
 /**
+ * Handle log messages posted from the utility process in the main process.
+ *
+ * See: [Note: Using Electron APIs in UtilityProcess]
+ *
+ * @param message The arbitrary message that was received as an argument to the
+ * "message" event invoked on a {@link UtilityProcess}.
+ *
+ * @returns true if the message was recognized and handled, and false otherwise.
+ */
+export const processUtilityProcessLogMessage = (
+    logTag: string,
+    message: unknown,
+) => {
+    const m = message; /* shorter alias */
+    if (m && typeof m == "object" && "method" in m && "p" in m) {
+        const p = m.p;
+        switch (m.method) {
+            case "log.errorString":
+                if (typeof p == "string") {
+                    log.error(`${logTag} ${p}`);
+                    return true;
+                }
+                break;
+            case "log.info":
+                if (Array.isArray(p)) {
+                    // Need to cast from any[] to unknown[]
+                    log.info(logTag, ...(p as unknown[]));
+                    return true;
+                }
+                break;
+            case "log.debugString":
+                if (typeof p == "string") {
+                    log.debug(() => `${logTag} ${p}`);
+                    return true;
+                }
+                break;
+            default:
+                break;
+        }
+    }
+    return false;
+};
+
+/**
  * Ente's logger.
  *
  * This is an object that provides three functions to log at the corresponding
